dcgan_tutorial/requirements.txt
absl-py==0.15.0
aiohttp==3.7.4.post0
astor==0.8.1
async-timeout==3.0.1
attrs==21.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
coverage==5.5
cryptography==3.4.7
cycler==0.11.0
Cython==0.29.24
gast==0.2.2
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.36.1
h5py==2.10.0
idna==3.3
idna-ssl==1.1.0
imageio==2.9.0
importlib-metadata==4.8.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
numpy==1.19.2
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
Pillow==8.3.1
pip==21.2.2
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==3.0.4
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.2
setuptools==58.0.4
six==1.16.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tornado==6.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==0.16.1
wheel==0.37.0
wincertstore==0.2
win-inet-pton==1.1.0
wrapt==1.12.1
yarl==1.6.3
zipp==3.6.0
